`XAS`: Move Pseudo Load to `setting.py` and clean code
Changes requested changes for PR #580:

* Moves the loading of pseudos and core wavefunction
data to `setting.py`. `setting.py` now collects labels
for pseudos and core_wfc_data, while `workchain.py`
loads the nodes from the labels.
* Replaces `set_panel_value()` content with call to
`self._update_element_select_panel()`.
* Fixes typos and removes redundant code noted in
latest review.
