+++ src/aiidalab_qe/plugins/xas/setting.py
@@ -2,12 +2,16 @@
 """Panel for XAS plugin.
 
 """
+import os
+import tarfile
 from importlib import resources
+from pathlib import Path
 
 import ipywidgets as ipw
+import requests
 import traitlets as tl
 import yaml
-from aiida.orm import StructureData
+from aiida import orm
 
 from aiidalab_qe.common.panel import Panel
 from aiidalab_qe.plugins import xas as xas_folder
@@ -16,11 +20,104 @@ PSEUDO_TOC = yaml.safe_load(resources.read_text(xas_folder, "pseudo_toc.yaml"))
 pseudo_data_dict = PSEUDO_TOC["pseudos"]
 xch_elements = PSEUDO_TOC["xas_xch_elements"]
 
+base_url = "https://github.com/PNOGillespie/Core_Level_Spectra_Pseudos/raw/main"
+head_path = f"{Path.home()}/.local/lib"
+dir_header = "cls_pseudos"
+functionals = ["pbe"]
+core_wfc_dir = "core_wfc_data"
+gipaw_dir = "gipaw_pseudos"
+ch_pseudo_dir = "ch_pseudos/star1s"
+
+
+def _load_or_import_nodes_from_filenames(in_dict, path, core_wfc_data=False):
+    for filename in in_dict.values():
+        try:
+            orm.load_node(filename)
+        except BaseException:
+            if not core_wfc_data:
+                new_upf = orm.UpfData(f"{path}/{filename}", filename=filename)
+                new_upf.label = filename
+                new_upf.store()
+            else:
+                new_singlefile = orm.SinglefileData(
+                    f"{path}/{filename}", filename="stdout"
+                )
+                new_singlefile.label = filename
+                new_singlefile.store()
+
+
+def _download_extract_pseudo_archive(func):
+    dir = f"{head_path}/{dir_header}/{func}"
+    archive_filename = f"{func}_ch_pseudos.tgz"
+    remote_archive_filename = f"{base_url}/{func}/{archive_filename}"
+    local_archive_filename = f"{dir}/{archive_filename}"
+
+    env = os.environ.copy()
+    env["PATH"] = f"{env['PATH']}:{Path.home() / '.local' / 'lib'}"
+
+    response = requests.get(remote_archive_filename, timeout=30)
+    response.raise_for_status()
+    with open(local_archive_filename, "wb") as handle:
+        handle.write(response.content)
+        handle.flush()
+        response.close()
+
+    with tarfile.open(local_archive_filename, "r:gz") as tarfil:
+        tarfil.extractall(dir)
+
+
+url = f"{base_url}"
+for func in functionals:
+    dir = f"{head_path}/{dir_header}/{func}"
+    os.makedirs(dir, exist_ok=True)
+    archive_filename = f"{func}_ch_pseudos.tgz"
+    archive_found = False
+    for entry in os.listdir(dir):
+        if entry == archive_filename:
+            archive_found = True
+    if not archive_found:
+        _download_extract_pseudo_archive(func)
+
+
+# Check all the pseudos/core-wfc data files in the TOC dictionary
+# and load/check all of them before proceeding. Note that this
+# approach relies on there not being multiple instances of nodes
+# with the same label.
+for func in functionals:
+    gipaw_pseudo_dict = pseudo_data_dict[func]["gipaw_pseudos"]
+    core_wfc_dict = pseudo_data_dict[func]["core_wavefunction_data"]
+    core_hole_pseudo_dict = pseudo_data_dict[func]["core_hole_pseudos"]
+    main_path = f"{head_path}/{dir_header}/{func}"
+    core_wfc_dir = f"{main_path}/core_wfc_data"
+    gipaw_dir = f"{main_path}/gipaw_pseudos"
+    ch_pseudo_dir = f"{main_path}/ch_pseudos/star1s"
+    # First, check that the local directories contain what's in the pseudo_toc
+    for pseudo_dir, pseudo_dict in zip(
+        [gipaw_dir, core_wfc_dir, ch_pseudo_dir],
+        [gipaw_pseudo_dict, core_wfc_dict, core_hole_pseudo_dict],
+    ):
+        pseudo_toc_mismatch = os.listdir(pseudo_dir) != pseudo_dict.values()
+
+    # Re-download the relevant archive if there is a mismatch
+    if pseudo_toc_mismatch:
+        _download_extract_pseudo_archive(func)
+
+    _load_or_import_nodes_from_filenames(
+        in_dict=gipaw_pseudo_dict,
+        path=gipaw_dir,
+    )
+    _load_or_import_nodes_from_filenames(
+        in_dict=core_wfc_dict, path=core_wfc_dir, core_wfc_data=True
+    )
+    _load_or_import_nodes_from_filenames(
+        in_dict=core_hole_pseudo_dict["1s"], path=ch_pseudo_dir
+    )
+
 
 class Setting(Panel):
     title = "XAS Settings"
     identifier = "xas"
-    input_structure = tl.Instance(StructureData, allow_none=True)
+    input_structure = tl.Instance(orm.StructureData, allow_none=True)
     protocol = tl.Unicode(allow_none=True)
 
     element_selection_title = ipw.HTML(
@@ -77,7 +174,7 @@ class Setting(Panel):
     def __init__(self, **kwargs):
         self.gipaw_pseudos = pseudo_data_dict["pbe"]["gipaw_pseudos"]
         self.core_hole_pseudos = pseudo_data_dict["pbe"]["core_hole_pseudos"]["1s"]
-        self.core_wfc_data = pseudo_data_dict["pbe"]["core_wavefunction_data"]
+        self.core_wfc_data_dict = pseudo_data_dict["pbe"]["core_wavefunction_data"]
 
         self.element_and_ch_treatment = ipw.VBox(layout=ipw.Layout(width="100%"))
 
@@ -114,8 +211,6 @@ class Setting(Panel):
         super().__init__(**kwargs)
 
     def get_panel_value(self):
-        """Return a dictionary with the input parameters for the plugin."""
-
         elements_list = []
         core_hole_treatments = {}
         for entry in self.element_and_ch_treatment.children:
@@ -125,39 +220,48 @@ class Setting(Panel):
                 elements_list.append(element)
                 core_hole_treatments[element] = ch_treatment
 
+        pseudo_labels = {}
+        core_wfc_data_labels = {}
+        for element in elements_list:
+            pseudo_labels[element] = {
+                "gipaw": self.gipaw_pseudos[element],
+                "core_hole": self.core_hole_pseudos[element],
+            }
+            core_wfc_data_labels[element] = self.core_wfc_data_dict[element]
+
         parameters = {
             "core_hole_treatments": core_hole_treatments,
             "elements_list": elements_list,
             # "structure_type": self.structure_type.value,
-            "gipaw_pseudo": self.gipaw_pseudos,
-            "ch_pseudo": self.core_hole_pseudos,
-            "core_wfc_data": self.core_wfc_data,
+            "pseudo_labels": pseudo_labels,
+            "core_wfc_data_labels": core_wfc_data_labels,
             "supercell_min_parameter": self.supercell_min_parameter.value,
         }
         return parameters
 
-    def set_panel_value(self, input_dict):
+    def set_panel_value(self):
         """Load a dictionary with the input parameters for the plugin."""
 
-        ch_pseudos = self.core_hole_pseudos
-        structure = self.input_structure
-        available_elements = [k for k in ch_pseudos]
-        elements_to_select = sorted(
-            [
-                kind.symbol
-                for kind in structure.kinds
-                if kind.symbol in available_elements
-            ]
-        )
+        # ch_pseudos = self.core_hole_pseudos
+        # structure = self.input_structure
+        # available_elements = [k for k in ch_pseudos]
+        # elements_to_select = sorted(
+        #     [
+        #         kind.symbol
+        #         for kind in structure.kinds
+        #         if kind.symbol in available_elements
+        #     ]
+        # )
 
-        element_and_ch_treatment_options = {}
-        for element in elements_to_select:
-            if element in xch_elements:
-                element_and_ch_treatment_options[element] = "xch_smear"
-            else:
-                element_and_ch_treatment_options[element] = "full"
+        # element_and_ch_treatment_options = {}
+        # for element in elements_to_select:
+        #     if element in xch_elements:
+        #         element_and_ch_treatment_options[element] = "xch_smear"
+        #     else:
+        #         element_and_ch_treatment_options[element] = "full"
 
-        self.element_and_ch_treatment_options = element_and_ch_treatment_options
+        # self.element_and_ch_treatment_options = element_and_ch_treatment_options
+        self._update_element_select_panel()
         # self.structure_type.value = input_dict.get("structure_type", "crystal")
 
     @tl.observe("input_structure")
